+++ tests/testthat/mod/env.r
@@ -1,0 +1,17 @@
+#' @export
+e1 = box::topenv()
+
+#' @export
+e2 = function () {
+    box::topenv()
+}
+
+#' @export
+e3 = function () {
+    e2()
+}
+
+#' @export
+e4 = function () {
+    function () box::topenv()
+}